Fixing some compiler warnings on comparison between signed and unsigned integer expressions. D.W. 20151215
git-svn-id: file:///Users/MacAir/Dropbox/xolotl-cleaned/branches/petscsolver_branch@980 72105f29-0fe1-43bd-9363-b1e281e962a2

+++ xolotl/xolotlCore/reactants/psiclusters/PSICluster.cpp
@@ -141,7 +141,7 @@ void PSICluster::createDissociationConnectivity() {
 	if (size == 1) {
 		// all the cluster of the same type dissociate into it
 		auto allSameTypeReactants = network->getAll(typeName);
-		for (int i = 0; i < allSameTypeReactants.size(); i++) {
+		for (unsigned int i = 0; i < allSameTypeReactants.size(); i++) {
 			auto cluster = (PSICluster *) allSameTypeReactants[i];
 			// X_1 cannot dissociate and X_2 --> X + X was already
 			// counted in the previous step
@@ -813,10 +813,10 @@ std::vector<int> PSICluster::getConnectivity() const {
 
 	// The reaction and dissociation vectors must have a length equal to the
 	// number of clusters
-	if (reactionConnVector.size() != connectivityLength) {
+	if (reactionConnVector.size() != (unsigned int)connectivityLength) {
 		throw std::string("The reaction vector is an incorrect length");
 	}
-	if (dissociationConnVector.size() != connectivityLength) {
+	if (dissociationConnVector.size() != (unsigned int)connectivityLength) {
 		throw std::string("The dissociation vector is an incorrect length");
 	}
